base_requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argilla==1.6.0
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
black==21.4b2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
dataclasses-json==0.5.7
Deprecated==1.2.13
dnspython==2.3.0
effdet==0.3.0
et-xmlfile==1.1.0
fastapi==0.95.1
filelock==3.11.0
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
future==0.18.3
fvcore==0.1.5.post20221221
google-auth==2.17.3
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.56.4
gptcache==0.1.10
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.53.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.13.4
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
iopath==0.1.9
Jinja2==3.1.2
joblib==1.2.0
jsons==1.6.3
kiwisolver==1.4.4
langchain==0.0.139
layoutparser==0.3.4
loguru==0.7.0
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.14.1
openai==0.27.4
openapi-schema-pydantic==1.2.4
opencv-python==4.6.0.66
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pathspec==0.11.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.9.0
Pillow==9.5.0
pinecone-client==2.2.1
portalocker==2.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.7
pydot==1.4.2
Pygments==2.15.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.0.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.26.1
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
tiktoken==0.3.3
timm==0.6.13
tokenizers==0.13.3
toml==0.10.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.28.0
typing-inspect==0.8.0
typing_extensions==4.5.0
typish==1.9.3
unstructured==0.5.2
unstructured-inference==0.2.7
urllib3==1.26.15
uvicorn==0.21.1
Wand==0.6.11
Werkzeug==2.2.3
wrapt==1.14.1
XlsxWriter==3.1.0
yacs==0.1.8
yarl==1.8.2
